bug fixed, npls were loaded as float

+++ main_master.c
@@ -25,7 +25,7 @@ int main(int argc, char *argv[]) {
     xmin = atof(argv[1]) - atof(argv[3]); 
     ymin = atof(argv[2]) - atof(argv[3]);
     radius = atof(argv[3]);
-    npls = atof(argv[4]);
+    npls = atoi(argv[4]);
     maxiter = atoi(argv[5]);
     char *outputfile = argv[6];
